Mueve las criterias al package modelo.criteria
