+++ src/main/java/sic/modelo/criteria/BusquedaFacturaCompraCriteria.java
@@ -1,10 +1,11 @@
-package sic.modelo;
+package sic.modelo.criteria;
 
 import java.util.Date;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import sic.modelo.TipoDeComprobante;
 
 @Data
 @AllArgsConstructor
