+++ src/main/java/sic/modelo/criteria/BusquedaPedidoCriteria.java
@@ -1,25 +1,30 @@
-package sic.modelo;
+package sic.modelo.criteria;
 
 import java.util.Date;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import sic.modelo.EstadoPedido;
+import sic.modelo.TipoDeEnvio;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class BusquedaCajaCriteria {
+public class BusquedaPedidoCriteria {
 
     private Date fechaDesde;
     private Date fechaHasta;
+    private Long idCliente;
+    private Long idUsuario;
+    private Long idViajante;
+    private Long nroPedido;
+    private EstadoPedido estadoPedido;
+    private TipoDeEnvio tipoDeEnvio;
+    private Long idProducto;
     private Long idEmpresa;
-    private int cantidadDeRegistros;
-    private Long idUsuarioApertura;
-    private Long idUsuarioCierre;
     private Integer pagina;
     private String ordenarPor;
     private String sentido;
-    
 }
