+++ src/main/java/sic/modelo/criteria/BusquedaReciboCriteria.java
@@ -1,9 +1,10 @@
-package sic.modelo;
+package sic.modelo.criteria;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import sic.modelo.Movimiento;
 
 import java.util.Date;
 
@@ -11,15 +12,19 @@ import java.util.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class BusquedaGastoCriteria {
+public class BusquedaReciboCriteria {
 
   private Date fechaDesde;
   private Date fechaHasta;
   private Long idEmpresa;
-  private Long idUsuario;
-  private Long idFormaDePago;
-  private Long nroGasto;
+  private Long numSerie;
+  private Long numRecibo;
   private String concepto;
+  private Long idCliente;
+  private Long idProveedor;
+  private Long idUsuario;
+  private Long idViajante;
+  private Movimiento movimiento;
   private Integer pagina;
   private String ordenarPor;
   private String sentido;
