+++ src/main/java/sic/modelo/criteria/BusquedaTransportistaCriteria.java
@@ -1,4 +1,4 @@
-package sic.modelo;
+package sic.modelo.criteria;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BusquedaTransportistaCriteria {
+
   private String nombre;
   private Long idProvincia;
   private Long idLocalidad;
