+++ src/main/java/sic/modelo/criteria/BusquedaUbicacionCriteria.java
@@ -1,4 +1,4 @@
-package sic.modelo;
+package sic.modelo.criteria;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
