+++ src/main/java/sic/service/IFacturaService.java
@@ -7,6 +7,8 @@ import sic.modelo.*;
 import java.util.List;
 
 import org.springframework.data.domain.Page;
+import sic.modelo.criteria.BusquedaFacturaCompraCriteria;
+import sic.modelo.criteria.BusquedaFacturaVentaCriteria;
 
 import javax.validation.Valid;
