+++ src/main/java/sic/service/IGastoService.java
@@ -5,7 +5,7 @@ import java.util.Date;
 import java.util.List;
 
 import org.springframework.data.domain.Page;
-import sic.modelo.BusquedaGastoCriteria;
+import sic.modelo.criteria.BusquedaGastoCriteria;
 import sic.modelo.Empresa;
 import sic.modelo.FormaDePago;
 import sic.modelo.Gasto;
