+++ src/main/java/sic/service/IProductoService.java
@@ -7,6 +7,7 @@ import sic.modelo.*;
 import java.util.List;
 import java.util.Map;
 import org.springframework.data.domain.Page;
+import sic.modelo.criteria.BusquedaProductoCriteria;
 import sic.modelo.dto.ProductosParaActualizarDTO;
 
 import javax.validation.Valid;
