+++ src/main/java/sic/service/IUbicacionService.java
@@ -2,6 +2,7 @@ package sic.service;
 
 import org.springframework.data.domain.Page;
 import sic.modelo.*;
+import sic.modelo.criteria.BusquedaLocalidadCriteria;
 
 import javax.validation.Valid;
 import java.util.List;
