+++ src/main/java/sic/service/impl/ProveedorServiceImpl.java
@@ -23,6 +23,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import sic.modelo.criteria.BusquedaProveedorCriteria;
 import sic.service.IProveedorService;
 import sic.exception.BusinessServiceException;
 import sic.repository.ProveedorRepository;
