+++ src/main/java/sic/service/impl/ReciboServiceImpl.java
@@ -29,6 +29,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
 import sic.modelo.*;
+import sic.modelo.criteria.BusquedaReciboCriteria;
 import sic.repository.ReciboRepository;
 import sic.service.*;
 import sic.exception.BusinessServiceException;
